consilidate archive into quality control step
