+++ tally_ho/apps/tally/templates/quality_control/print_cover.html
